fix: prevent uint8_t overflow in AGC attack and recovery guard conditions (#145)
Cast gain_step_down + min_gain and agc_base_gain + gain_step_up to uint16_t
before comparison to avoid silent wrap-around when sums exceed 255.

With aggressive AGC values (e.g. step=200, min=100), the attack-path sum
wraps to 44, the guard fires when it should not, and the AGC reduces gain
below min_gain, potentially driving the LNA into an unsafe low-gain state.

The recovery path has the same overflow risk: (250+50) & 0xFF = 44 <= max_gain
passes when it should be clamped, allowing gain to exceed the configured ceiling.

Added two adversarial regression tests that verify correct clamping with
extreme gain values that trigger the overflow boundary.

+++ 9_Firmware/9_1_Microcontroller/9_1_1_C_Cpp_Libraries/ADAR1000_AGC.cpp
@@ -44,7 +44,7 @@ void ADAR1000_AGC::update(bool fpga_saturation)
         saturation_event_count++;
         holdoff_counter = 0;
 
-        if (agc_base_gain >= gain_step_down + min_gain) {
+        if ((uint16_t)agc_base_gain >= (uint16_t)gain_step_down + (uint16_t)min_gain) {
             agc_base_gain -= gain_step_down;
         } else {
             agc_base_gain = min_gain;
@@ -60,7 +60,7 @@ void ADAR1000_AGC::update(bool fpga_saturation)
         if (holdoff_counter >= holdoff_frames) {
             holdoff_counter = 0;
 
-            if (agc_base_gain + gain_step_up <= max_gain) {
+            if ((uint16_t)agc_base_gain + (uint16_t)gain_step_up <= (uint16_t)max_gain) {
                 agc_base_gain += gain_step_up;
             } else {
                 agc_base_gain = max_gain;
